+++ src/useSQLite.ts
@@ -17,6 +17,19 @@ export type SQLiteProps = {
  */
 export interface SQLiteHook extends AvailableResult {
     /**
+     * Init the web store
+     * @returns Promise<void>
+     * @since 2.1.0
+     */
+    initWebStore(): Promise<void>;
+    /**
+     * Save the datbase to the web store
+     * @param database
+     * @returns Promise<void>
+     * @since 2.1.0
+     */
+    saveToStore(database: string): Promise<void>;
+     /**
      * Echo a value
      * @param value
      * @returns Promise<{value: string}>
@@ -197,10 +210,7 @@ export interface Result {
 /**
  * useSQLite Hook
  */
-export const useSQLite = ({
-    onProgressImport,
-    onProgressExport
-}: SQLiteProps): SQLiteHook  => {
+export const useSQLite = (onProgress? : SQLiteProps): SQLiteHook  => {
     const platform = Capacitor.getPlatform();
     const sqlitePlugin: any = CapacitorSQLite;
     const mSQLite = useMemo(() => {
@@ -212,16 +222,20 @@ export const useSQLite = ({
         let importListener: any = null;
         let exportListener: any = null;    
         if(platform != "electron") {   
-            if(onProgressImport && sqlitePlugin) importListener =
-                sqlitePlugin.addListener('sqliteImportProgressEvent',
-                (e: any) => {
-                    onProgressImport(e.progress);
-                });
-            if(onProgressExport && sqlitePlugin) exportListener =
-                sqlitePlugin.addListener('sqliteExportProgressEvent',
-                (e: any) => {
-                    onProgressExport(e.progress);
-                });
+            if( onProgress ) { 
+                if(onProgress.onProgressImport && sqlitePlugin) importListener =
+                    sqlitePlugin.addListener('sqliteImportProgressEvent',
+                    (e: any) => {
+                        if(typeof onProgress.onProgressImport !== 'undefined')
+                        onProgress.onProgressImport(e.progress);
+                    });
+                if(onProgress.onProgressExport && sqlitePlugin) exportListener =
+                    sqlitePlugin.addListener('sqliteExportProgressEvent',
+                    (e: any) => {
+                        if(typeof onProgress.onProgressExport !== 'undefined')
+                        onProgress.onProgressExport(e.progress);
+                    });
+                }
             }
         return () => {
             if(platform != "electron") {   
@@ -235,6 +249,41 @@ export const useSQLite = ({
         useSQLite: isFeatureAvailable('CapacitorSQLite', 'useSQLite')
     }
 
+    /**
+     * Initialize the Web Store
+     */
+    const initWebStore = useCallback(async (): Promise<void> => {
+        if(platform != "web") { 
+            return Promise.reject(`Not implemented on platform ${platform}`);
+        }
+
+        try {
+            await mSQLite.initWebStore();
+            return Promise.resolve();
+        } catch (err) {
+            return Promise.reject(err);
+        }
+    }, [mSQLite]);
+    /**
+     * Save the Database to store
+     * @param dbName string
+     */
+    const saveToStore = useCallback(async (dbName: string): Promise<void> => {
+        if(platform != "web") { 
+            return Promise.reject(`Not implemented on platform ${platform}`);
+        }
+        if(dbName.length > 0) {
+            try {
+                await mSQLite.saveToStore(dbName);
+                return Promise.resolve();
+            } catch (err) {
+                return Promise.reject(err);
+            }
+        } else {
+            return Promise.reject('Must provide a database name');
+        }
+    }, [mSQLite]);
+    
 
     const echo = useCallback(async (value: string): Promise<any> => {
         if(value) {
@@ -581,6 +630,8 @@ export const useSQLite = ({
 
     if (!availableFeaturesN.useSQLite) {
         return {
+            initWebStore: featureNotAvailableError,
+            saveToStore: featureNotAvailableError,
             echo: featureNotAvailableError,
             getPlatform: featureNotAvailableError,
             getCapacitorSQLite: featureNotAvailableError,
@@ -611,7 +662,7 @@ export const useSQLite = ({
             isConnection, isDatabase, getDatabaseList, addSQLiteSuffix,
             deleteOldDatabases, checkConnectionsConsistency, 
             isSecretStored, setEncryptionSecret, changeEncryptionSecret,
-            isAvailable: true};
+            initWebStore, saveToStore, isAvailable: true};
     }
 
 }